backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
scikit-learn==1.7.1
joblib
google-generativeai
Pillow
# SpeechRecognition
gTTS
python-dotenv
python-multipart
gunicorn